==> notebooks/step_1_data_pre_processing.ipynb <==
import pandas as pd
import numpy as np
import os
from pathlib import Path

notebook_path = os.path.join(os.getcwd())
data_path = Path(notebook_path).parent.joinpath('data')

# Read 2016 and 2017 worksheets data. 2016 contains all the data back to 1878.
df_2017 = pd.read_excel(data_path / 'Fires_100.xlsx')
df_2016 = pd.read_excel(data_path / 'Fires_100.xlsx',1)

# Remove missing data
df_2017.rename(columns={'YEAR_':'YEAR'}, inplace=True)
df_2017.YEAR.fillna(2017, inplace=True)
df_2016 = df_2016[df_2016.YEAR!='<Null>'].dropna(subset='YEAR')

# Concat 2016 and 2017 data
df = pd.concat([df_2016, df_2017])

# Remove unused columns
df = df[['YEAR', 'AGENCY', 'UNIT_ID', 'FIRE_NAME', 'INC_NUM',
       'ALARM_DATE', 'CONT_DATE', 'CAUSE', 'REPORT_AC',
       'GIS_ACRES', 'C_METHOD', 'OBJECTIVE']]

# Rows with 'null' string should be treated like real nulls
df = df.replace('<Null>', np.nan)

# Save
df.to_csv(data_path / 'wildfires_preprocessed.csv', index=False)
