==> notebooks/step_3_task_1.ipynb <==
import pandas as pd
import numpy as np
import requests
import plotly.express as px
import os
from pathlib import Path

notebook_path = os.path.join(os.getcwd())
data_path = Path(notebook_path).parent.joinpath('data')

# Read data
df = pd.read_csv(data_path / 'wildfires_with_geocoord.csv', parse_dates=True)

# Make grouping by unit and year - counts
grouped_count = df.groupby(['unit', 'lat', 'lon', 'year']).count()
grouped_count = grouped_count.reset_index()
grouped_count = grouped_count[['unit', 'lat', 'lon', 'year','agency']]
grouped_count = grouped_count.rename(columns={'agency':'count'})

grouped_area = df.groupby(['unit', 'lat', 'lon', 'year']).sum()
grouped_area = grouped_area.reset_index()
grouped_area = grouped_area[['unit', 'lat', 'lon', 'year', 'gis_acres']]
grouped_area = grouped_area.rename(columns={'gis_acres':'area_burned'})

grouped = pd.concat([grouped_count, grouped_area[['area_burned']]], axis=1)

grouped = grouped.sort_values('year')
grouped = grouped[grouped.year<2018]
# Save
grouped.to_csv(data_path / 'wildfires_grouped_task1.csv', index=False)
