==> notebooks/step_4_task_2.ipynb <==
import pandas as pd
import numpy as np
import requests
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import os
from pathlib import Path

notebook_path = os.path.join(os.getcwd())
data_path = Path(notebook_path).parent.joinpath('data')

# Read data
df = pd.read_csv(data_path / 'wildfires_with_geocoord.csv', parse_dates=True)

# Make grouping by year - counts
grouped_counts = df.groupby('year').count()
grouped_counts = grouped_counts.reset_index()
grouped_counts = grouped_counts[['year', 'agency']]
grouped_counts = grouped_counts.rename(columns={'agency':'count'})
grouped_counts = grouped_counts.sort_values('year')

# Make grouping by year - acres burned
grouped_area = df.groupby('year').sum()
grouped_area = grouped_area.reset_index()
grouped_area = grouped_area[['year', 'gis_acres']]
grouped_area = grouped_area.rename(columns={'gis_acres':'area_burned'})
grouped_area = grouped_area.sort_values('year')

# Merge
df_grouped = grouped_counts.merge(grouped_area)
# Save
df_grouped.to_csv(data_path / 'wildfires_grouped_task2.csv', index=False)
