==> notebooks/step_6_task_4.ipynb <==
import pandas as pd
import numpy as np
import requests
import plotly.express as px
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import os
from pathlib import Path

notebook_path = os.path.join(os.getcwd())
data_path = Path(notebook_path).parent.joinpath('data')

# Read data
df = pd.read_csv(data_path / 'wildfires_with_geocoord.csv', parse_dates=True)

# some data wrangling
df['month'] = pd.to_datetime(df.alarm_date).dt.month
df.dropna(subset='month', inplace=True)
grouped = df.groupby(['month', 'cause']).count()
grouped.reset_index(inplace=True)
grouped = grouped[['month', 'cause', 'agency']]
grouped.columns = ['month', 'cause', 'count']
grouped
grouped.to_csv(data_path / 'wildfires_grouped_task4_2.csv', index=False)

# 4.2
df_4_2 = pd.read_csv(data_path / 'wildfires_grouped_task4_2.csv')
order = df_4_2.groupby('cause').sum().sort_values('count', ascending=False)
order_list = order.index.to_list()

# 4.1
df_4_1 = pd.read_csv(data_path / 'wildfires_grouped_task4_2.csv')
df_4_1 = df_4_2.groupby('month').sum()
df_4_1.reset_index(inplace=True)
df_4_1.to_csv(data_path / 'wildfires_grouped_task4_1.csv', index=False)


# 4.3
df.dropna(subset='month', inplace=True)
grouped = df.groupby(['month', 'year']).count()
grouped.reset_index(inplace=True)
grouped = grouped[['month', 'year', 'agency']]
grouped.columns = ['month', 'year', 'count']
grouped.sort_values('year', inplace=True)
df_12_months = pd.DataFrame(columns=['month'], data={'month':np.arange(1,13)})
df_years = pd.DataFrame(columns=['year'], data={'year':np.arange(1898,2019)})
temp = df_years.merge(df_12_months, how='cross')
df_merged = temp.merge(grouped, how='outer', left_on=['year','month'], right_on=['year','month']).sort_values(['year', 'month'])
df_merged.fillna(0, inplace=True)
df_merged.to_csv(data_path / 'wildfires_grouped_task4_3.csv', index=False)
